refactor(config-ui): adjust the type for bp detail page
